progress

+++ tp2/validador.h
@@ -2,8 +2,9 @@
 #define __VALIDADOR_H__
 
 #include <stdbool.h>
+#include <stdlib.h>
 #include "comando.h"
 
-comando_t validar_comando(int argc, char** argv);
+comando_t validar_comando(size_t argc, char** argv);
 
 #endif
